src/requirements.txt
requests==2.9.1
lxml==3.6.0
pystache==0.5.4
cssselect==0.9.1
pillow==3.2.0